+++ changelog.md
@@ -1,5 +1,10 @@
 V1.0
-Fully functional Lego Minotaurus game
+-Fully functional Lego Minotaurus game.
 
 V1.1
-Simplyfication of mapTools module and replacement of board class with grid variable
+-Simplyfication of mapTools module and replacement of board class with grid variable.
+
+V1.5
+-Some methods in controller class cascaded into subordinate classes (figure, walls etc).
+-pygame initiation functions and several map drawing methods migrated into mapTools.py
+-Bug fixes